Move analytics logic to analytics util
